--- recommender/README.md
@@ -1,10 +1,0 @@
-# simple recommendation system
-
-- data: [movielens 1M](https://grouplens.org/datasets/movielens/1m/)
-
-### item-based
-
-- Item-based collaborative filtering using adjusted cosine similarity
-  |버전|RMSE|소요시간|
-  |--:|:--:|:--:|
-  |ver.1|1.042|4.5m ~ 8m|
